add fitroutines

+++ C++/MainFrame.h
@@ -47,7 +47,7 @@ class MainFrame: public wxFrame {
 
   bool dataloaded;
 
-  int *dataleft, *dataright;
+  double *dataleft, *dataright;
 
   enum {ID_sliderFramenum=wxID_HIGHEST + 1, ID_sliderThres, ID_sliderLeft, \
 	ID_sliderRight};
